b/e test refactoring
