+++ src/test/java/codesquad/user/UserDtoTest.java
@@ -1,6 +1,7 @@
-package codesquad.domain;
+package codesquad.user;
 
 
+import codesquad.user.dto.UserDto;
 import org.junit.Before;
 import org.junit.Test;
